+++ pos_mobile/doc/changelog.rst
@@ -1,3 +1,10 @@
+`1.1.0`
+-------
+
+- **NEW:** Payment screen interface
+- **NEW:** Quantity indicator of products added to the current order
+- **NEW:** Order list scrolling
+
 `1.0.0`
 -------